Rename to in-memory bus.
