+++ config/buses/in-memory/tmp/bus2.yaml
@@ -2,13 +2,13 @@
 apiVersion: eventing.knative.dev/v1alpha1
 kind: Channel
 metadata:
-  name: foo-bus
+  name: foo
 spec:
   provisioner:
     ref:
       apiVersion: eventing.knative.dev/v1alpha1
       kind: ClusterProvisioner
-      name: stub-bus-provisioner
+      name: in-memory-bus-provisioner
   channelable:
     subscribers:
       - sinkableDomain: message-dumper-foo.default.svc.cluster.local # Already exists
@@ -20,14 +20,15 @@ spec:
 apiVersion: eventing.knative.dev/v1alpha1
 kind: Channel
 metadata:
-  name: bar-bus
+  name: bar
 spec:
   provisioner:
     ref:
       apiVersion: eventing.knative.dev/v1alpha1
       kind: ClusterProvisioner
-      name: stub-bus-provisioner
+      name: in-memory-bus-provisioner
   channelable:
     subscribers:
       - sinkableDomain: message-dumper-bar.default.svc.cluster.local # Already exists
         callableDomain: event-changer.default.svc.cluster.local
+
