+++ pkg/controller/eventing/inmemory/channel/controller.go
@@ -14,11 +14,13 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package clusterprovisioner
+package channel
 
 import (
 	eventingv1alpha1 "github.com/knative/eventing/pkg/apis/eventing/v1alpha1"
+	"github.com/knative/eventing/pkg/system"
 	"go.uber.org/zap"
+	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/controller"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
@@ -28,18 +30,25 @@ import (
 const (
 	// controllerAgentName is the string used by this controller to identify
 	// itself when creating events.
-	controllerAgentName = "stub-bus-cluster-provisioner-controller"
+	controllerAgentName = "in-memory-bus-channel-controller"
+	ConfigMapName       = "in-memory-bus-dispatcher-config-map"
+)
+
+var (
+	DefaultConfigMapKey = types.NamespacedName{
+		Namespace: system.Namespace,
+		Name:      ConfigMapName,
+	}
 )
 
 // ProvideController returns a flow controller.
 func ProvideController(mgr manager.Manager, logger *zap.Logger) (controller.Controller, error) {
-	logger = logger.With(zap.String("controller", controllerAgentName))
-
-	// Setup a new controller to Reconcile ClusterProvisioners that are Stub buses.
-	r :=  &reconciler{
-		mgr: mgr,
-		recorder: mgr.GetRecorder(controllerAgentName),
-		logger: logger,
+	// Setup a new controller to Reconcile Channels that belong to this Cluster Provisioner
+	// (in-memory buses).
+	r := &reconciler{
+		configMapKey: DefaultConfigMapKey,
+		recorder:     mgr.GetRecorder(controllerAgentName),
+		logger:       logger,
 	}
 	c, err := controller.New(controllerAgentName, mgr, controller.Options{
 		Reconciler: r,
@@ -49,17 +58,17 @@ func ProvideController(mgr manager.Manager, logger *zap.Logger) (controller.Cont
 		return nil, err
 	}
 
-	// Watch ClusterProvisioners.
+	// Watch Channels.
 	err = c.Watch(&source.Kind{
-		Type: &eventingv1alpha1.ClusterProvisioner{},
+		Type: &eventingv1alpha1.Channel{},
 	}, &handler.EnqueueRequestForObject{})
 	if err != nil {
-		logger.Error("Unable to watch ClusterProvisioners.", zap.Error(err), zap.Any("type", &eventingv1alpha1.ClusterProvisioner{}))
+		logger.Error("Unable to watch Channels.", zap.Error(err), zap.Any("type", &eventingv1alpha1.Channel{}))
 		return nil, err
 	}
 
-	// TODO: Should we watch the K8s service as well? If it changes, we probably should change it
-	// back.
+	// TODO: Should we watch the K8s service and Istio Virtual Service as well? If they change, we
+	// probably should change it back.
 
 	return c, nil
 }
